add button component and styling
