+++ components/MainColor/styles.js
@@ -1,4 +1,12 @@
 import styled from "styled-components";
+import { Roboto } from "@next/font/google";
+
+const roboto = Roboto({
+  weight: ["100", "900"],
+  style: ["normal", "italic"],
+  subsets: ["latin"],
+  display: "swap",
+});
 
 export const StyledLabel = styled.label`
   margin: 25px 0;
@@ -10,22 +18,15 @@ export const StyledInput = styled.input`
   height: 100px;
   width: 33%;
   max-width: 200px;
+  background-color: var(--background-color);
 `;
 
 export const StyledDiv = styled.div`
   margin-top: 20px;
+  background-color: var(--background-color);
 `;
 
-export const StyledButton = styled.button`
-  border: none;
-  border-radius: 100px;
-  margin: 20px;
-  text-decoration: none;
-  height: 40px;
-  width: 120px;
-  font-size: 20px;
-  background-color: #e6e6e6;
-`;
 export const StyledForm = styled.form`
   text-align: center;
+  background-color: var(--background-color);
 `;
